+++ docs/changes.rst
@@ -10,6 +10,7 @@ Not yet released.
 * Display nearby keys on monolingual translations.
 * Added support for groupping string shapings.
 * Recommend upgrade to new Weblate version in the system checks.
+* Provide more detailed analysis for duplicate language alert.
 
 Weblate 3.10.3
 --------------